+++ ExtrasKit/Swift/Array+ExtrasKit.swift
@@ -43,6 +43,33 @@ extension ExtrasKitWrapper where Base: RangeReplaceableCollection, Base.Element:
 
         return (nil, base)
     }
+
+    /// Returns a new collection of the same type containing, in order, the
+    /// elements of the original collection that remove all duplicate elements.
+    ///
+    ///     var cast = ["Vivien", "Marlon", "Kim", "Karl", "Vivien"]
+    ///     let deduplication = cast.ek.deduplicate { $0 }
+    ///     print(deduplication)
+    ///     // Prints "["Vivien", "Marlon", "Kim", "Karl"]"
+    ///
+    /// - Parameter filter: A closure that takes an element of the
+    ///   sequence as its argument and returns returns a transformed value indicating
+    ///   whether the element is duplicated and original element should be included
+    ///   in the returned collection.
+    /// - Returns: A collection of the elements after removing all duplicate elements.
+    ///
+    /// - Complexity: O(*n * n*), where *n* is the length of the collection.
+    public mutating func deduplicate<E: Equatable>(_ filter: (Base.Element) -> E) -> [Base.Element] {
+
+        return base.reduce([]) { unique, element in
+            let key = filter(element)
+            if !unique.map({ filter($0) }).contains(key) {
+                return unique + [element]
+            }
+
+            return unique
+        }
+    }
 }
 
 extension ExtrasKitWrapper where Base: Sequence, Base.Element: AdditiveArithmetic {